extract visibleTodoAtoms hook
