+++ examples/todo/src/ui/TodoApp/MainSection/TodoList/TodoList.tsx
@@ -1,22 +1,13 @@
 import * as React from 'react'
-import { useAtomSlice } from 'klyva'
 import { TodoItem } from './TodoItem'
-import { useFilterValue, useTodoListAtom } from '../../../../bridge'
+import { useVisibleTodoAtoms } from '../../../../bridge'
 
 export const TodoList = () => {
-  const filter = useFilterValue()
-  const todoListAtom = useTodoListAtom()
-  const visibleTodoAtoms = useAtomSlice(
-    todoListAtom,
-    ({ checked }) =>
-      filter === 'all' ||
-      (filter === 'completed' && checked) ||
-      (filter === 'active' && !checked),
-  )
+  const visibleTodoAtoms = useVisibleTodoAtoms()
   return (
     <ul className="todo-list">
       {visibleTodoAtoms.map((todoAtom, index) => (
-        <TodoItem key={filter + index} todoAtom={todoAtom} />
+        <TodoItem key={index} todoAtom={todoAtom} />
       ))}
     </ul>
   )
